requirements.txt
tensorflow>=2.10.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0
PySide6>=6.0.0
Pillow>=8.3.0

# API Integration Dependencies
google-cloud-videointelligence>=2.0.0
boto3>=1.26.0
requests>=2.28.0

# Advanced Video Analysis Dependencies
ultralytics>=8.0.0  # YOLOv8, YOLOv9, YOLO11 support
transformers>=4.20.0  # BLIP, BLIP-2, Qwen2-VL, LLaVA support
mediapipe>=0.10.0  # OpenPose alternative
torch>=1.12.0
torchvision>=0.13.0
matplotlib>=3.5.0
pillow>=9.0.0  # Required for image processing in transformers

# GitHub Models Integration
azure-ai-inference>=1.0.0b9
python-dotenv>=0.19.0